app/pyproject.toml
[project]
name = "app1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.16.0",
    "pandas>=3.0.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
